Add TravianLoader and TravianProxy
Use TravianLoader to inject a dependency and allow to replace the
implementation used to fetch data. This way it is very easy to test
the Updater and Loader modules and the rake tasks which use them
without ever needing to issue requests to an external source.

TravianProxy therefore is the local and fake implementation for the
Travian gem allowing local manipulation to the data fetched to be used
in a test and development environment.

The implementation actually being used is configured in the current
environment using the travian_loader_source config variable.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>