Major overhaul of the api and add proper specs
Add default format json to api scope module

Add Api::V1::BaseController, make all Api::V1::Controllers inherit from
it and add an after_filter to set the content-type to reflect the
version being used followed by the json format

Correct bug in ApiVersion#matches? where it whould raise an error when
no header was present.

Add tests to all routes and test for the present of the correct
content-type in the response.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>